// src/saga/domain/saga.enums.ts
export enum SagaStatus {
  ACTIVE = "ACTIVE",
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
  COMPENSATING = "COMPENSATING",
  COMPENSATED = "COMPENSATED",
  MANUAL_INTERVENTION_REQUIRED = "MANUAL_INTERVENTION_REQUIRED",
  CANCELLED = "CANCELLED",
}

export enum SagaStep {
  ORDER_PREPARATION = "ORDER_PREPARATION",
  BUDGET_REQUESTED = "BUDGET_REQUESTED",
  WAITING_BUDGET_APPROVAL = "WAITING_BUDGET_APPROVAL",
  WAITING_PAYMENT = "WAITING_PAYMENT",
  STOCK_RESERVATION_REQUESTED = "STOCK_RESERVATION_REQUESTED",
  STOCK_RESERVED = "STOCK_RESERVED",
  EXECUTION_REQUESTED = "EXECUTION_REQUESTED",
  IN_EXECUTION = "IN_EXECUTION",
  FINISHED = "FINISHED",
}

export enum CompensationStatus {
  NOT_REQUIRED = "NOT_REQUIRED",
  PENDING = "PENDING",
  IN_PROGRESS = "IN_PROGRESS",
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
}

export enum CompensationActionStatus {
  NOT_REQUIRED = "NOT_REQUIRED",
  PENDING = "PENDING",
  REQUESTED = "REQUESTED",
  WAITING_RESOURCE_CONFIRMATION = "WAITING_RESOURCE_CONFIRMATION",
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
}

export type NextCompensationAction =
  "STOCK_RELEASE" | "PAYMENT_REFUND" | "COMPLETE" | "NONE";